Extract method to get available games (headed-client) to a utility class 'AvailableGamesFileSystemReader'
